feat(power-consumption): move formatting to separate file
