+++ src/lib/components/profile/FactoryDetails.svelte
@@ -8,44 +8,13 @@
 	import alerts from '@/lib/stores/alerts.svelte';
 	import { stage } from '@/lib/stores/stage.svelte';
 	import { ZapIcon } from '@lucide/svelte';
+	import { formatPower } from '@/lib/calculator/unit';
 
 	let inputs = $state<{ item: Item; amount: number }[]>([]),
 		outputs = $state<{ item: Item; amount: number }[]>([]);
 
 	let powerConsumption = $state('0');
 	let powerUnit = $state('W');
-
-	function formatPower(
-		watts: number,
-		decimals: number = 2,
-	): { value: string; unit: string; rawNumber: number } {
-		if (watts === 0) return { value: '0', unit: 'W', rawNumber: 0 };
-
-		const userLocale = typeof navigator !== 'undefined' ? navigator.language : 'en-US';
-
-		const units = [
-			{ value: 1e12, label: 'terawatt', short: 'TW' },
-			{ value: 1e9, label: 'gigawatt', short: 'GW' },
-			{ value: 1e6, label: 'megawatt', short: 'MW' },
-			{ value: 1e3, label: 'kilowatt', short: 'kW' },
-			{ value: 1, label: 'watt', short: 'W' },
-		];
-
-		const absWatts = Math.abs(watts);
-		const match = units.find(u => absWatts >= u.value) || units[units.length - 1];
-		const calculatedValue = watts / match.value;
-
-		const formattedValue = new Intl.NumberFormat(userLocale, {
-			maximumFractionDigits: decimals,
-			minimumFractionDigits: 0,
-		}).format(calculatedValue);
-
-		return {
-			value: formattedValue,
-			unit: match.short,
-			rawNumber: calculatedValue,
-		};
-	}
 
 	function calculate() {
 		if (!active.profile) return;
